refactor: Reorganize editor module into extensions/ and commands/ subdirectories
- Rename extensions.ts → extensions-client.ts for clarity
- Move all extension files to lib/domain/editor/extensions/
  - wiki-link-node.ts → extensions/wiki-link.ts
  - callout-extension.ts → extensions/callout.ts
  - tag-extension.ts → extensions/tag.ts
  - task-list-input-rule.ts → extensions/task-list.ts
  - bullet-list-backspace.ts → extensions/bullet-list.ts
  - Plus suggestion files
- Move command files to lib/domain/editor/commands/
  - slash-commands.tsx, slash-commands-menu.tsx
- Update imports in extensions-client.ts and MarkdownEditor.tsx

Improves module organization and makes extension types more discoverable.
