+++ components/content/editor/MarkdownEditor.tsx
@@ -13,7 +13,7 @@
 
 import { useEditor, EditorContent } from "@tiptap/react";
 import { useEffect, useRef, useState } from "react";
-import { getEditorExtensions } from "@/lib/domain/editor/extensions";
+import { getEditorExtensions } from "@/lib/domain/editor/extensions-client";
 import type { JSONContent } from "@tiptap/core";
 import { LinkDialog } from "./LinkDialog";
 import { BubbleMenu } from "./BubbleMenu";
