+++ lib/domain/editor/extensions/wiki-link.ts
